initial code pushed
This work gives an example of compact hyperbolic 3-manifolds that are strongly isospectral but have nonisomorphic rational cohomology rings.

The file AlmostConjugateExample shows that the chosen group has almost conjugate subgroups. The file SunadaPair implements the Sunada method. In the file CupProduct we give a method for computing the cup product map for aspherical spaces and show that our examples are strongly isospectral but have nonisomorphic rational cohomology rings.